Cache is updated on launch; improved error handling.

+++ src/app.py
@@ -13,6 +13,7 @@ grib2.noaa_params = ['maxt', 'mint', 'wspd', 'wdir']
 grib2.geodata_path = '/usr/local/share/degrib/geodata'
 
 mutex = Lock()
+downloading_mutex = Lock()
 download_base = '/home/nebain/ndfd/auto/'
 download_dir = download_base + '/active'
 
@@ -45,24 +46,38 @@ def forecast(lat, lon, params):
 
 @app.route('/update_cache')
 def update_cache():
+	new_download_dir = get_new_download_dir()
+	lock_acquired = downloading_mutex.acquire(False)
+
+	if not lock_acquired:
+		print "Service Unavailable"
+		abort(503) # Service Unavailable
+
 	try:
-		new_download_dir = get_new_download_dir()
 		new_files = grib2.download(download_dir, new_download_dir)
 		if new_files:
 			old_download_dir = os.path.realpath(download_dir)
-			mutex.acquire()
-			if os.path.exists(download_dir) and os.path.islink(download_dir):
-				os.unlink(download_dir)
-			elif os.path.exists(download_dir) and os.path.isdir(download_dir):
-				shutil.rmtree(download_dir)
-			os.symlink(new_download_dir, download_dir)
-			mutex.release()
+			with mutex:
+				if os.path.exists(download_dir) and os.path.islink(download_dir):
+					os.unlink(download_dir)
+				elif os.path.exists(download_dir) and os.path.isdir(download_dir):
+					shutil.rmtree(download_dir)
+				os.symlink(new_download_dir, download_dir)
 			if os.path.exists(old_download_dir) and not os.path.islink(old_download_dir):
 				shutil.rmtree(old_download_dir)
+
+		downloading_mutex.release()
 	except:
+		downloading_mutex.release()
 		abort(500)
-		raise
-	return 'Success'
+
+	if new_files:
+		return ('Success - new files downloaded.', 200)
+	else:
+		return ('Success - no update needed.', 304)
 
 if __name__ == '__main__':
+	print "Updating local NDFD file cache..."
+	new_download_dir = get_new_download_dir()
+	new_files = grib2.download(download_dir, new_download_dir)
 	app.run(host='0.0.0.0', threaded=True)#, port=80)
